Split format module into submodules
