+++ src/format/mod.rs
@@ -1,0 +1,23 @@
+//! Utility module to convert data to types and back by
+//! specific formats like: JSON, BSON, TOML, YAML, XML.
+//!
+//! All types here are lazy and it's necessary to
+//! use `Into` and `From` traits to get (convert) the data.
+
+use failure::Error;
+
+pub mod nothing;
+pub mod json;
+
+pub use self::nothing::Nothing;
+pub use self::json::Json;
+
+/// A representation of a value which can be stored and restored as a text.
+pub type Text = Result<String, Error>;
+
+/// A representation of a value which can be stored and restored as a binary.
+pub type Binary = Result<Vec<u8>, Error>;
+
+/// A helper which represents a specific format.
+#[doc(hidden)]
+pub type Format<T> = Result<T, Error>;
