+++ src/format/nothing.rs
@@ -1,0 +1,31 @@
+//! Contains an implementation of empty serialization format (`Nothing`).
+
+use failure::err_msg;
+use super::{Text, Binary};
+
+/// A representation of an empty data. Nothing stored. Nothing restored.
+pub struct Nothing;
+
+impl Into<Text> for Nothing {
+    fn into(self) -> Text {
+        Err(err_msg("nothing"))
+    }
+}
+
+impl From<Text> for Nothing {
+    fn from(_: Text) -> Nothing {
+        Nothing
+    }
+}
+
+impl Into<Binary> for Nothing {
+    fn into(self) -> Binary {
+        Err(err_msg("nothing"))
+    }
+}
+
+impl From<Binary> for Nothing {
+    fn from(_: Binary) -> Nothing {
+        Nothing
+    }
+}
